[finish] finish room_page function
