+++ project/backend/app.js
@@ -1,6 +1,5 @@
 const express = require('express');
 require('express-async-errors');
-
 require('module-alias/register');
 
 const bodyParser = require('body-parser');
@@ -10,18 +9,27 @@ const dotenv = require('dotenv');
 dotenv.config();
 
 const app = express();
-const PORT = process.env.PORT;
+const http = require("http");
+const { Server } = require("socket.io");
+
+const server = http.createServer(app);
+const io = new Server(server, {
+  cors: { origin: "http://localhost:5173" },
+});
+
+const PORT = process.env.PORT || 3000;
 
 app.use(bodyParser.json());
-
 app.use(cors({
-    origin: ["http://localhost:5173", "http://127.0.0.1:5173"]
-    // origin: "*"
+  origin: ["http://localhost:5173", "http://127.0.0.1:5173"]
 }));
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+require("@/socket_services/Room/socket")(io);
 
 
-// routes are here
-// user's inventory aka "my sets page"
+// User's inventory aka "my sets page"
 const folderRoutes = require("@/routes/User_Inventory/folderRoutes.js");
 const setRoutes = require("@/routes/User_Inventory/setRoutes.js");
 const wordRoutes = require("@/routes/User_Inventory/wordRoutes.js");
@@ -30,7 +38,7 @@ app.use("/folders", folderRoutes);
 app.use("/sets", setRoutes);
 app.use("/words", wordRoutes);
 
-// search add aka "main landing page"
+// Search add aka "main landing page"
 const landing_userRoutes = require('@/routes/SearchAdd/userRoutes.js');
 const landing_wordRoutes = require('@/routes/SearchAdd/wordRoutes.js');
 const landing_setRoutes = require('@/routes/SearchAdd/setRoutes.js');
@@ -39,23 +47,22 @@ app.use('/api/users', landing_userRoutes);
 app.use('/api/words', landing_wordRoutes);
 app.use('/api/set', landing_setRoutes);
 
-// account related aka "log in page"
+// Account related aka "log in page"
 const authRoutes = require('@/routes/Account/authRoutes.js');
-
 app.use('/auth', authRoutes);
 
-// routes end
-
 app.get("/test", (req, res) => {
-    res.status(200).send({ 'success': 'success' });
-})
-
-// globally checker for error handling, so we dont need catch for any async func in backend
-app.use((err, req, res, next) => {
-    console.log(err);
-    res.status(err.status || 500).send("something is wrong...\n detected in global error handler");
+  res.status(200).send({ 'success': 'success' });
 });
 
-app.listen(PORT, () => {
-    console.log(`Server running on http://localhost:${PORT}`);
+const room_setRoutes = require('@/routes/Room/setRoutes.js');
+app.use(room_setRoutes);
+
+app.use((err, req, res, next) => {
+  console.log(err);
+  res.status(err.status || 500).send("something is wrong...\n detected in global error handler");
+});
+
+server.listen(PORT, () => {
+  console.log(`Server running on http://localhost:${PORT}`);
 });
