+++ project/backend/src/routes/Room/chatRoutes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const chatController = require("@/controllers/chatController");
+const chatController = require("@/controllers/Room/chatController");
 
 
 router.post("/message", chatController.SaveMessage);
