+++ project/frontend/package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "axios": "^1.7.7",
     "pinia": "^2.3.0",
+    "socket.io": "^4.8.1",
+    "socket.io-client": "^4.8.1",
     "vue": "^3.5.12",
     "vue-router": "^4.4.5"
   },
