+++ project/frontend/src/router/index.js
@@ -5,6 +5,7 @@ import HomeLoggedIn from '@/views/HomeLoggedIn.vue';
 import Home from '@/views/Home.vue';
 import VerifyEmail from '@/views/Page_Account/VerifyEmail.vue';
 import ResetPassword from '@/views/Page_Account/ResetPassword.vue';
+import Room from '@/views/Page_Room/Room.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -38,6 +39,11 @@ const router = createRouter({
       path: '/reset-password',
       name: 'ResetPassword',
       component: ResetPassword,
+    },
+    { 
+      path: '/Room',
+      name: 'Room',
+      component: Room,
     }
   ],
 })
