--- room/backend/.env.example
@@ -1,7 +1,0 @@
-PORT = 3000 or your port
-
-MYSQL_HOST = 'localhost (or tour IP addr)'
-MYSQL_USER = 'root (or your custom name)'
-MYSQL_PW = 'your pw'
-MYSQL_DB = 'vocab_let_u_learn'
-
