--- room/backend/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
-    "dev": "nodemon app.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "_moduleAliases": {
-    "@": "src"
-  },
-  "dependencies": {
-    "dotenv": "^16.4.7",
-    "express": "^4.21.2",
-    "module-alias": "^2.2.3",
-    "mysql2": "^3.11.5",
-    "redis": "^4.7.0",
-    "socket.io": "^4.8.1"
-  }
-}
