--- room/backend/src/db.js
@@ -1,24 +1,0 @@
-const mysql = require('mysql2/promise');
-const dotenv = require('dotenv');
-dotenv.config();
-
-const pool = mysql.createPool({
-    host: process.env.MYSQL_HOST,
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PW,
-    database: 'vocab_let_u_learn'
-});
-
-// testing connection
-async function testConnection() {
-    try {
-        const connection = await pool.getConnection();
-        console.log('Database connection successful');
-        connection.release(); // Release the connection back to the pool
-    } catch (error) {
-        console.error('Database connection failed:', error.message);
-    }
-}
-testConnection();
-
-module.exports = pool;
