--- room/frontend/.env.example
@@ -1,1 +1,0 @@
-VITE_API_BASE_URL = 'your backend base URL'
