--- room/frontend/.vscode/extensions.json
@@ -1,6 +1,0 @@
-{
-  "recommendations": [
-    "Vue.volar",
-    "vitest.explorer"
-  ]
-}
