--- room/frontend/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "frontend",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "test:unit": "vitest"
-  },
-  "dependencies": {
-    "axios": "^1.7.7",
-    "ioredis": "^5.4.1",
-    "module-alias": "^2.2.3",
-    "socket.io-client": "^4.8.1",
-    "vue": "^3.5.12",
-    "vue-router": "^4.4.5"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^5.1.4",
-    "@vue/test-utils": "^2.4.6",
-    "jsdom": "^25.0.1",
-    "vite": "^5.4.10",
-    "vite-plugin-vue-devtools": "^7.5.4",
-    "vitest": "^2.1.8"
-  }
-}
