--- room/frontend/src/App.vue
@@ -1,19 +1,0 @@
-<script>
-import { RouterLink, RouterView } from 'vue-router'
-
-export default {
-  name: 'App',
-};
-</script>
-
-<template>
-  <RouterView />
-</template>
-
-<style>
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-</style>
