--- room/frontend/src/assets/style.css
@@ -1,8 +1,0 @@
-/* src/assets/styles.css */
-html, body {
-    height: 100%;
-    width: 100%;
-    margin: 0;
-    padding: 0;
-    overflow-x: hidden; /* 防止水平滾動條 */
-}
