--- room/frontend/src/router/index.js
@@ -1,22 +1,0 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import Room from '@/views/Room.vue';
-
-
-const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    
-      {
-        path: '/',
-        redirect: '/Room',
-      },
-      
-      { 
-        path: '/Room',
-        name: 'Room',
-        component: Room,
-    }
-  ],
-})
-
-export default router
