--- room/frontend/src/services/APIclient.js
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-const apiClient = axios.create({
-    baseURL: import.meta.env.VITE_API_BASE_URL,
-    headers: {
-        'Content-Type': 'application/json',
-    },
-    timeout: 10000
-});
-
-export default apiClient;
